+++ casestudy/app/Http/Controllers/UserController.php
@@ -23,9 +23,9 @@ class UserController extends Controller
     }
 
     public function create() {
-//        if (!$this->userCan('admin')) {
-//            abort(403);
-//        }
+        if (!$this->userCan('admin')) {
+            abort(403);
+        }
         $roles = Role::all();
         return view('admin.users.create', compact('roles'));
     }
@@ -37,9 +37,9 @@ class UserController extends Controller
     }
 
     public function destroy($id){
-//        if (!$this->userCan('admin')) {
-//            abort(403);
-//        }
+        if (!$this->userCan('admin')) {
+            abort(403);
+        }
         $user = $this->userService->findById($id);
         $user->delete();
         $message = 'Successfully Deleted The User!';
@@ -47,9 +47,9 @@ class UserController extends Controller
     }
 
     public function edit($id) {
-//        if (!$this->userCan('admin')) {
-//            abort(403);
-//        }
+        if (!$this->userCan('admin')) {
+            abort(403);
+        }
         $roles = Role::all();
         $users = $this->userService->findById($id);
         return view('admin.users.edit', compact('users', 'roles'));
